splitting modules
