+++ packages/sw-background-sync-queue/build.js
@@ -59,9 +59,9 @@ module.exports = () => {
     }),
     buildJSBundle({
       rollupConfig: {
-        entry: path.join(__dirname, 'src', 'lib', 'queue.js'),
+        entry: path.join(__dirname, 'src', 'lib', 'request-queue.js'),
         format: 'umd',
-        moduleName: 'goog.backgroundSyncQueue.test',
+        moduleName: 'goog.backgroundSyncQueue.test.requestQueue',
         plugins: [
           resolve({
             jsnext: true,
@@ -75,7 +75,28 @@ module.exports = () => {
           commonjs(),
         ],
       },
-      outputName: 'build/test/queue.js',
+      outputName: 'build/test/request-queue.js',
+      projectDir: __dirname,
+    }),
+    buildJSBundle({
+      rollupConfig: {
+        entry: path.join(__dirname, 'src', 'lib', 'constants.js'),
+        format: 'umd',
+        moduleName: 'goog.backgroundSyncQueue.test.constants',
+        plugins: [
+          resolve({
+            jsnext: true,
+            main: true,
+            browser: true,
+          }),
+          rollupBabel({
+            plugins: ['transform-async-to-generator', 'external-helpers'],
+            exclude: 'node_modules/**',
+          }),
+          commonjs(),
+        ],
+      },
+      outputName: 'build/test/constants.js',
       projectDir: __dirname,
     }),
   ]);
