+++ packages/sw-background-sync-queue/src/lib/background-sync-queue.js
@@ -1,10 +1,10 @@
 import RequestManager from './request-manager';
-import Queue from './queue';
+import RequestQueue from './request-queue';
 import {maxAge} from './constants';
 
 class BackgroundSyncQueue {
 	constructor({config, callbacks, queueName}) {
-		this._queue = new Queue(
+		this._queue = new RequestQueue(
 			Object.assign({}, {maxAge: maxAge}, config),
 			queueName);
 		this._requestManager = new RequestManager({callbacks, queue: this._queue});
