+++ packages/sw-background-sync-queue/src/lib/request-manager.js
@@ -1,5 +1,5 @@
 import {putResponse} from './response-manager';
-
+import {getQueueableRequest, getFetchableRequest} from './queue-utils';
 /**
  * Class to handle all the request related
  * transformations, replaying, event handling
@@ -70,56 +70,4 @@ class RequestManager {
 	}
 }
 
-/**
- * takes a request and gives back JSON object that is storable in IDB
- *
- * @param {Request} request request object to transform
- * into iDB storable object
- * @param {Object} config config object to be
- * stored along in the iDB
- * @return {Object} indexable object for iDB
- *
- * @memberOf RequestManager
- */
-async function getQueueableRequest(request, config) {
-	let requestObject={
-		config: config,
-		metadata: {
-			creationTimestamp: Date.now(),
-		},
-	};
-	let headerObject = {};
-	request.headers.forEach((value, field)=> {
-		headerObject[field] = value;
-	});
-	requestObject.request = {
-		url: request.url,
-		headers: headerObject,
-		mode: request.mode,
-		method: request.method,
-		redirect: request.redirect,
-	};
-	let requestBody = await request.text();
-	if (requestBody.length > 0) {
-		requestObject.request.body = requestBody;
-	}
-	return requestObject;
-}
-
-async function getFetchableRequest(idbRequestObject) {
-	let reqObject = {
-		mode: idbRequestObject.mode,
-		method: idbRequestObject.method,
-		redirect: idbRequestObject.redirect,
-		headers: new Headers(idbRequestObject.headers),
-	};
-	if(idbRequestObject.body) {
-		reqObject.body = idbRequestObject.body;
-	}
-	return new Request(idbRequestObject.url, reqObject);
-}
-
 export default RequestManager;
-export {
-	getQueueableRequest,
-};
