+++ packages/sw-background-sync-queue/test/unit/index.html
@@ -39,7 +39,8 @@
     <script src="/node_modules/sw-testing-helpers/browser/sw-utils.js"></script>
     <script src="/node_modules/mockdate/src/mockdate.js"></script>
     <script src="/node_modules/sinon/pkg/sinon.js"></script>
-    <script src="/packages/sw-background-sync-queue/build/test/queue.js"></script>
+    <script src="/packages/sw-background-sync-queue/build/test/constants.js"></script>
+    <script src="/packages/sw-background-sync-queue/build/test/request-queue.js"></script>
     <!--
       Timeout is extended to ensure tests for max-cache-age
       have enough time to complete
